Verbesserung von Kommentaren und Debug-Output, bessere Implementierung der Logik zur Bestimmung des verantwortlichen Peers für ein Hash-Value

+++ peer.h
@@ -7,7 +7,7 @@
 
 typedef struct {
     UT_hash_handle hh;
-    uint16_t key;
+    int key;
     int fd;
     crud_packet* request;
 } client_info;
